fix(python): split tests FIPS from generic tests
